Refactoring code to meet current React standarts
