+++ client/src/App.js
@@ -5,6 +5,10 @@ import { useState } from "react";
 
 function App() {
   const [score, setScore] = useState({ Blue: 0, Red: 0 });
+
+  /*This function is passed as a callback to the Equation component. If the answer given to the 
+  equation is the correct answer, the Equation component will pass 'true' to gotRightAnswer, if the answer is wrong it will pass 'false'. 
+  That will be used to update the score state*/
   function gotRightAnswer(isAnswerRight) {
     if (isAnswerRight) {
       setScore((current) => {
