+++ client/src/components/helperFunctions.js
